raftstore: adjust code
